Remove Paketo-specific BOM from top-level package
