+++ build_test.go
@@ -579,9 +579,6 @@ api = "0.6"
 
 	context("when there are bom entries in the build metadata", func() {
 		it("persists a build.toml", func() {
-			algorithm512, err := packit.GetBOMChecksumAlgorithm("sha512")
-			Expect(err).ToNot(HaveOccurred())
-
 			packit.Build(func(ctx packit.BuildContext) (packit.BuildResult, error) {
 				return packit.BuildResult{
 					Build: packit.BuildMetadata{
@@ -591,18 +588,8 @@ api = "0.6"
 							},
 							{
 								Name: "another-example",
-								Metadata: packit.BOMMetadata{
-									Version: "0.5",
-									Checksum: packit.BOMChecksum{
-										Algorithm: packit.SHA256,
-										Hash:      "12345",
-									},
-									Source: packit.BOMSource{
-										Checksum: packit.BOMChecksum{
-											Algorithm: algorithm512,
-											Hash:      "some-source-sha",
-										},
-									},
+								Metadata: map[string]string{
+									"some-key": "some-value",
 								},
 							},
 						},
@@ -619,14 +606,7 @@ api = "0.6"
 				[[bom]]
 					name = "another-example"
 				[bom.metadata]
-					version = "0.5"
-				[bom.metadata.checksum]
-					algorithm = "SHA-256"
-					hash = "12345"
-				[bom.metadata.source]
-					[bom.metadata.source.checksum]
-						algorithm = "SHA-512"
-						hash = "some-source-sha"
+					some-key = "some-value"
 			`))
 		})
 
@@ -654,8 +634,8 @@ api = "0.4"
 								},
 								{
 									Name: "another-example",
-									Metadata: packit.BOMMetadata{
-										Version: "0.5",
+									Metadata: map[string]string{
+										"some-key": "some-value",
 									},
 								},
 							},
@@ -801,8 +781,8 @@ api = "0.6"
 							},
 							{
 								Name: "another-example",
-								Metadata: packit.BOMMetadata{
-									Version: "0.5",
+								Metadata: map[string]string{
+									"some-key": "some-value",
 								},
 							},
 						},
@@ -819,7 +799,7 @@ api = "0.6"
 				[[bom]]
 					name = "another-example"
 				[bom.metadata]
-					version = "0.5"
+					some-key = "some-value"
 			`))
 		})
 	})
@@ -1384,13 +1364,6 @@ api = "0.4"
 					}, packit.WithArgs([]string{binaryPath, layersDir, platformDir, planPath}), packit.WithExitHandler(exitHandler))
 					Expect(exitHandler.ErrorCall.Receives.Error).To(MatchError(ContainSubstring("permission denied")))
 				})
-			})
-		})
-
-		context("when the attempted BOM checksum algorithm is not supported", func() {
-			it("persists a build.toml", func() {
-				_, err := packit.GetBOMChecksumAlgorithm("RANDOM-ALG")
-				Expect(err).To(MatchError("failed to get supported BOM checksum algorithm: RANDOM-ALG is not valid"))
 			})
 		})
